move role_man into core, not optional extension (fix #130)
